Proposed fix: solution to the problem of wrong line number reports when compiling with jake (as requested by mrcarlberg)
`gcc -E -x c -P` was used in CFHTTPRequest.js to preprocess the input.
-P strips the line number information (along with lots of whitespace).
acorn.js therefore calculated wrong line numbers.

With the gcc -P option removed from CFHTTPRequest.js,
gcc emits source position lines in the form
'# lineNumber "file or other location"'.
The gcc command now includes a sed command to convert
the # at the beginning of a line into //# to be valid JS.
The case that a previous lines ends with a backslash -
escaping the newline - is handled by gcc - it joins the lines.
I changed getLineInfo() in acorn.js to interpret said line information
and correctly calculate the source line number.

Caveats:
I did not yet change the second path which is used when the locations
option is on (curLineLoc() / nextLineStart()) but would do so if
this solution is welcome.

I tested (jake test and some samples) on OS X 10.8.2.
Since cygwin is required under Windows, I expect sed to be
available there too.

Since this is my first foray into the Cappuccino source,
I may be missing a lot of things. Please double check.
There is another -P option in compiler.js and one in
./Tests/DetectMissingImports.js - I didn't touch these yet.

I changed some code I touched anyway to adhere to Contributions.md.
I kept the previous (2 space) top-level indentation to avoid a large
whitespace or jiggly code.